chore: clean up terminal component folder structure
