+++ apps/codeimage/src/App.tsx
@@ -7,6 +7,9 @@ import {frame$, setScale} from '@codeimage/store/frame';
 import {connectStoreWithQueryParams} from '@codeimage/store/plugins/connect-store-with-query-params';
 import {terminal$} from '@codeimage/store/terminal';
 import {Box, LoadingOverlay, PortalHost, SnackbarHost} from '@codeimage/ui';
+import {fromObservableObject} from '@core/hooks/from-observable-object';
+import {useModality} from '@core/hooks/isMobile';
+import {useEffects} from '@core/store/use-effect';
 import {initEffects} from '@ngneat/effects';
 import {createEffect, createSignal, on, Show} from 'solid-js';
 import {BottomBar} from './components/BottomBar/BottomBar';
@@ -20,12 +23,9 @@ import {SidebarPopoverHost} from './components/PropertyEditor/SidebarPopoverHost
 import {Canvas} from './components/Scaffold/Canvas/Canvas';
 import {Scaffold} from './components/Scaffold/Scaffold';
 import {Sidebar} from './components/Scaffold/Sidebar/Sidebar';
-import {DynamicTerminal} from './components/Terminal/dynamic/DynamicTerminal';
+import {DynamicTerminal} from './components/Terminal/DynamicTerminal/DynamicTerminal';
 import {ThemeSwitcher} from './components/ThemeSwitcher/ThemeSwitcher';
 import {Toolbar} from './components/Toolbar/Toolbar';
-import {fromObservableObject} from './core/hooks/from-observable-object';
-import {useModality} from './core/hooks/isMobile';
-import {useEffects} from './core/store/use-effect';
 import {uiStore} from './state/ui';
 import './theme/global.css';
