+++ apps/codeimage/src/components/Terminal/DynamicTerminal/DynamicTerminal.tsx
