+++ apps/codeimage/src/components/Terminal/GlassReflection/TerminalGlassReflection.tsx
@@ -1,0 +1,10 @@
+import {Box} from '@codeimage/ui';
+import * as styles from './TerminalGlassReflection.css';
+
+export const TerminalGlassReflection = () => {
+  return (
+    <Box class={styles.wrapper}>
+      <Box class={styles.content} />
+    </Box>
+  );
+};
