+++ apps/codeimage/src/components/Terminal/MacOsTerminal/MacOsTerminal.css.ts
