+++ apps/codeimage/src/components/Terminal/Tabs/Tab/Tab.css.ts
@@ -1,7 +1,7 @@
 import {themeVars} from '@codeimage/ui';
 import {createTheme, style} from '@vanilla-extract/css';
 import {recipe} from '@vanilla-extract/recipes';
-import {terminalVars} from '../terminal.css';
+import {terminalVars} from '../../terminal.css';
 
 export const [tabTheme, tabVars] = createTheme({
   tabHeight: '30px',
