+++ apps/codeimage/src/components/Terminal/Tabs/Tab/WindowTab.tsx
@@ -5,10 +5,10 @@ import {createSortable} from '@thisbeyond/solid-dnd';
 import {assignInlineVars} from '@vanilla-extract/dynamic';
 import {createMemo, createSignal, onMount, Show, VoidProps} from 'solid-js';
 import {exportExclude as _exportExclude} from '@core/directives/exportExclude';
-import {CloseIcon} from '../../Icons/CloseIcon';
+import {CloseIcon} from '../../../Icons/CloseIcon';
 import * as styles from './Tab.css';
-import {TabIcon} from './TabIcon/TabIcon';
-import {TabName} from './TabName/TabName';
+import {TabIcon} from '../TabIcon/TabIcon';
+import {TabName} from '../TabName/TabName';
 
 const exportExclude = _exportExclude;
