+++ apps/codeimage/src/components/Terminal/Tabs/TabAddButton/TabAddButton.tsx
