+++ apps/codeimage/src/components/Terminal/Tabs/TabIcon/TabIcon.tsx
