--- apps/codeimage/src/components/Terminal/TerminalGlassReflection.tsx
@@ -1,10 +1,0 @@
-import {Box} from '@codeimage/ui';
-import * as styles from './terminal.css';
-
-export const TerminalGlassReflection = () => {
-  return (
-    <Box class={styles.glassReflection.wrapper}>
-      <Box class={styles.glassReflection.content} />
-    </Box>
-  );
-};
