+++ apps/codeimage/src/components/Terminal/WindowsTerminal/WindowsTerminalControls.tsx
